requirements.txt
agent-framework-github-copilot
azure-storage-blob>=12.19.0